+++ x/nftmint/client/cli/query_minting_info_test.go
@@ -154,7 +154,7 @@ func TestListMintingInfo(t *testing.T) {
 		var resp types.QueryMintingInfosResponse
 		require.NoError(t, net.Config.Codec.UnmarshalJSON(out.Bytes(), &resp))
 		require.NoError(t, err)
-		require.Equal(t, len(objs), int(resp.Pagination.Total))
+		require.Equal(t, uint64(len(objs)), resp.Pagination.Total)
 		require.ElementsMatch(t,
 			nullify.Fill(objs),
 			nullify.Fill(resp.MintingInfo),
